+++ packages/platform_network_discovery/lib/platform_network_discovery.dart
@@ -1,3 +1,4 @@
 library;
 
 export 'src/local_discovery_models.dart';
+export 'src/discovery_service.dart';